Drop cgroupfs support
Drop support for directly modifying the blkio cgroup controller
parameters. There were/are no known users to this functionality. Plus we
only supported cgroups v1.
