+++ pkg/blockio/oci.go
@@ -20,8 +20,6 @@ import (
 	"fmt"
 
 	oci "github.com/opencontainers/runtime-spec/specs-go"
-
-	"github.com/intel/goresctrl/pkg/cgroups"
 )
 
 // OciLinuxBlockIO returns OCI LinuxBlockIO structure corresponding to the class.
@@ -43,7 +41,7 @@ func OciLinuxBlockIO(class string) (*oci.LinuxBlockIO, error) {
 	return &ociBlockio, nil
 }
 
-func ociLinuxWeightDevices(dws cgroups.DeviceWeights) []oci.LinuxWeightDevice {
+func ociLinuxWeightDevices(dws DeviceWeights) []oci.LinuxWeightDevice {
 	if len(dws) == 0 {
 		return nil
 	}
@@ -57,7 +55,7 @@ func ociLinuxWeightDevices(dws cgroups.DeviceWeights) []oci.LinuxWeightDevice {
 	return olwds
 }
 
-func ociLinuxThrottleDevices(drs cgroups.DeviceRates) []oci.LinuxThrottleDevice {
+func ociLinuxThrottleDevices(drs DeviceRates) []oci.LinuxThrottleDevice {
 	if len(drs) == 0 {
 		return nil
 	}
