--- pkg/cgroups/cgroupid_other.go
@@ -1,8 +1,0 @@
-//go:build !linux
-// +build !linux
-
-package cgroups
-
-func getID(path string) uint64 {
-	return 0
-}
